\name{wts}
\alias{wts}
\docType{data}
\title{Weights vector}
\description{This is a single column vector for weighting the importance of point locations.}
\usage{data(wts)}
\format{A single column vector of numeric values.}
\details{The weights can be specified according to any reasonable criteria specified by the user}
\source{This demonstration data has been manufactured for illustrative purposes only.}
\examples{
  data(wts)
  str(wts)
  plot(wts)
}
\keyword{datasets}
