\name{aspace-package}
\alias{aspace-package}
\alias{aspace}
\docType{package}
\title{A collection of functions for estimating centrographic statistics and computational geometries for spatial point patterns}
\description{A collection of functions for computing centrographic statistics (e.g., standard distance, standard deviation ellipse, standard deviation box) for observations taken at point locations. The 'aspace' package was originally conceived to aid in the analysis of spatial patterns of travel behaviour (see Buliung and Remmel, 2008).}
\details{
\tabular{ll}{
Package: \tab aspace\cr
Type: \tab Package\cr
Version: \tab 4.1.0\cr
Date: \tab 2023-09-05\cr
License: \tab GPL-3\cr
}
}
\author{Randy Bui, Ron N. Buliung, Tarmo K. Remmel}
\references{
  Bachi, R. 1963. Standard distance measures and related methods for spatial analysis. 
  Papers of the Regional Science Association 10: 83-132.   
  \cr\cr
  Buliung, R.N. and Remmel, T. (2008) Open source, spatial analysis, and activity travel 
  behaviour research: capabilities of the aspace package. Journal of 
  Geographical Systems, 10: 191-216.
  \cr\cr
  Buliung, R.N. and Kanaroglou, P.S. (2006) Urban form and household activity-travel 
  behaviour. Growth and Change, 37: 174-201.  
  \cr\cr
  Ebdon, D. 1988. Statistics in Geography 2nd Edition. Oxford UK: Blackwell. 
  \cr\cr
  Levine, N. 2002. CrimeStat II: A Spatial Statistics Program for the Analysis 
  of Crime Incident Locations (version 2.0) Houston TX/National Institute of Justice, 
  Washington DC: Ned Levine & Associates.
}
\keyword{arith}
